fix: download works for Macro WorkFlow type
This is still incomplete as when the download has finished working the isDone is still not set as true, as a result when the download has finished instead of displaying "done" in the specific row in the download column an empty string will be displayed.

Additional cleanup of the code.