feature: if the job is Macro check that there is a user script uploaded
If there is no script uploaded then the Start job item will not appear in the context menu. Currently the method responsible for doing this calls the "getFileContents" method which is slow. This will be replaced with a check to job.upload == true and a check that the user has uploaded a DEFAULT_MACRO_FILE file.