rename Benchmark.fxml to HPCWorkflow.fxml
This file was missed during refactoring and had to be renamed.

Also made a small modification to the TreeView in order to display only the columns that exist and resize them when the table is resized.

This is to conserve screen space and reduce user's effort to resize the columns when the window is resized.