fix: error output is now working for macro
The error output tab was labeled Snakemake output in the past and was therefore disabled for Macro. Later it was re-titled as "Error" tab for the macro but the actual functionality was not re-enabled. Now it is enabled and works. The error tab will be enabled when there are errors to display. Also renamed the "snakemakeOutputControl" to "logViewControl" as it is now used for displaying errors for the Macro Workflow type as well.