feat: Add macro scatter functionality
Added conversion from array to string and back again.
Changed the calls to include the name of the new refactored package.
return id of task added
added warning in code that is appended to the user's script
This comments mark the section of code that is automatically appended to the user script as well as warn the user not to remove this code section.