feat: add support for user's script that is named by the user
Instead of always using "user.ijm" the user will be able to choose any
*.ijm file of that the user likes.

When creating a new job there is a directory or file chooser depending
on the workflow type. If a user changes workflow type after filling in a
directory there will now be a warning if workflow is macro and there is
no selected file, if the opposite happens the file will be ignored. The
initially selected folder in file chooser and file picker is also chosen
correctly.
simplify code