fix: stop reset of directory filed in create job, remove file instead
When creating a new job there is a directory or file chooser depending on the workflow type. If a user changes workflow type after filling in a directory there will now be a warning if workflow is macro and there is no selected file, if the opposite happens the file will be ignored. The initially selected folder in file chooser and file picker is also chosen correctly.
simplify code 


remove unnecessary check 


clean code