remove always on top from progress dialog, run all javafx on its thread
Some parts of the code that should run on JavaFX thread did not run on the thread. This is fixed.

Moved some of the runOnFxThread from the Window class to the Controller class to make it easier to avoid mistakenly running JavaFX code on a non-JavaFx thread.

Removed the always on top property from the progress dialog as the modality of the windows works and the case where the dashboard is in front of the dialog is less annoying than having the progress dialog in front of all windows of the OS (including eclipse).