fix progress exception on close
The constructor and the "done" method are called by different threads in "HPCWorkflowControl" and sometimes the "Platform.runLater()" (used to run the GUI operations in the JavaFX thread) is called in the reverse order.

When this happens the constructor does not yet exist and an exception is produced.

The controller was created in the JavaFX thread while the done method needed the controller while running in the main thread. Sometimes the JavaFX thread was slower and the controller was still NULL at the time of closing the progress bar.