if the logs do not exit and the job is finished stop trying to download
The progress logs may not exist as the job may not report progress or it may have not been started yet. 

If the job does not report progress the logs will never be created causing the unnecessary checks until the progress window is closed. 

This fixes the issue. If the job has finished in any way it will stop trying to get the progress files if they do not exist at all.